Rename openai-speech to openai
