+++ packages/providers/openai/src/OpenAITranscriber.test.ts
