+++ packages/providers/openai/src/codec.test.ts
