+++ packages/providers/openai/src/codec.ts
