+++ packages/providers/openai/src/models.ts
