+++ recipes/package.json
@@ -12,7 +12,7 @@
     "@effect-uai/elevenlabs": "workspace:*",
     "@effect-uai/google": "workspace:*",
     "@effect-uai/jina": "workspace:*",
-    "@effect-uai/openai-speech": "workspace:*",
+    "@effect-uai/openai": "workspace:*",
     "@effect-uai/responses": "workspace:*",
     "effect": "4.0.0-beta.57"
   },
